Revert "Merge branch 'master' of https://github.com/InfiniTimeOrg/InfiniTime"
This reverts commit 2a819b654e0007e53a7e33796783d869000645a7, reversing
changes made to 93d099e944e501703f5859812a160672a98190a3.

+++ src/drivers/Hrs3300.h
@@ -30,8 +30,8 @@ namespace Pinetime {
       void Init();
       void Enable();
       void Disable();
-      uint32_t ReadHrs();
-      uint32_t ReadAls();
+      uint16_t ReadHrs();
+      uint16_t ReadAls();
       void SetGain(uint8_t gain);
       void SetDrive(uint8_t drive);
